+++ Pipfile
@@ -4,6 +4,8 @@ verify_ssl = true
 name = "pypi"
 
 [packages]
+pick = "*"
+pysimplegui = "*"
 
 [dev-packages]
